Refactor App into components
