+++ landingdodi/src/components/About.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function About() {
+  return (
+    <section id="about" className="about">
+      <h2>About DoDi</h2>
+      <p>
+        DoDi helps you manage documentation with ease. Our platform is built
+        for reliability and scalability.
+      </p>
+    </section>
+  );
+}
+
+export default About;
