+++ landingdodi/src/components/Footer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Footer() {
+  return (
+    <footer className="footer">
+      <p>&copy; {new Date().getFullYear()} DoDi. All rights reserved.</p>
+    </footer>
+  );
+}
+
+export default Footer;
