+++ landingdodi/src/components/Hero.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+function Hero() {
+  return (
+    <header className="hero">
+      <div className="hero-content">
+        <h1>Welcome to DoDi</h1>
+        <p>Your digital documentation solution</p>
+        <a className="cta-button" href="#contact">Get Started</a>
+      </div>
+    </header>
+  );
+}
+
+export default Hero;
