+++ landingdodi/src/components/OfertaEducativa.jsx
@@ -1,0 +1,25 @@
+import React from 'react';
+
+function OfertaEducativa() {
+  return (
+    <section id="features" className="features">
+      <h2>Features</h2>
+      <div className="feature-grid">
+        <div className="feature-item">
+          <h3>Automation</h3>
+          <p>Automate your document workflows.</p>
+        </div>
+        <div className="feature-item">
+          <h3>Integration</h3>
+          <p>Connect with popular tools.</p>
+        </div>
+        <div className="feature-item">
+          <h3>Security</h3>
+          <p>Keep your data safe.</p>
+        </div>
+      </div>
+    </section>
+  );
+}
+
+export default OfertaEducativa;
